+++ spec/puppetfile_spec.rb
@@ -163,5 +163,20 @@ RSpec.describe PuppetfileEditor::Puppetfile do
         'rabbitmq' => { new: '7.0.0', type: :git },
       )
     end
+
+    it 'compares two Puppetfiles with modules of different types' do
+      pf = described_class.new(File.join(fixtures_dir, 'compare', 'source.Puppetfile'))
+      pf.load
+
+      pf_new = described_class.new(File.join(fixtures_dir, 'compare', 'new_different_types.Puppetfile'))
+      pf_new.load
+
+      expect(pf.compare_with(pf_new, compare_across_types: true)).to eq(
+        'apache'   => { old: '2.0.0', new: '2.1.0', type: :hg },
+        'apt'      => { old: '4.1.0', new: '4.3.0', type: :forge },
+        'nginx'    => { old: '0.7.0', new: '0.7.1', type: :git },
+        'rabbitmq' => { new: '7.0.0', type: :git },
+      )
+    end
   end
 end